+++ buildSrc/src/main/kotlin/com/orange/ods/gradle/Dependencies.kt
@@ -20,6 +20,8 @@ object Dependencies {
     const val androidGradlePlugin = "com.android.tools.build:gradle:${Versions.androidGradlePlugin}"
     const val appCompat = "androidx.appcompat:appcompat:${Versions.appCompat}"
     const val browser = "androidx.browser:browser:${Versions.browser}"
+    const val coil = "io.coil-kt:coil:${Versions.coil}"
+    const val coilCompose = "io.coil-kt:coil-compose:${Versions.coil}"
     const val composeMaterial = "androidx.compose.material:material:${Versions.compose}"
     const val composeUi = "androidx.compose.ui:ui:${Versions.compose}"
     const val composeUiTooling = "androidx.compose.ui:ui-tooling:${Versions.compose}"
@@ -47,5 +49,6 @@ object Dependencies {
     const val material = "com.google.android.material:material:${Versions.material}"
     const val navigationCompose = "androidx.navigation:navigation-compose:${Versions.navigation}"
     const val testExtJUnit = "androidx.test.ext:junit:${Versions.testExtJUnit}"
+    const val timber = "com.jakewharton.timber:timber:${Versions.timber}"
     const val webkit = "androidx.webkit:webkit:${Versions.webkit}"
 }