+++ demo/src/main/java/com/orange/ods/demo/data/DataModule.kt
@@ -8,14 +8,11 @@
  * /
  */
 
-package com.orange.ods.demo.domain
+package com.orange.ods.demo.data
 
 import android.content.Context
-import com.orange.ods.demo.domain.datastore.DataStoreService
-import com.orange.ods.demo.domain.datastore.DataStoreServiceImpl
+import com.orange.ods.demo.data.recipes.RecipesRepositoryImpl
 import com.orange.ods.demo.domain.recipes.RecipesRepository
-import com.orange.ods.demo.domain.recipes.RecipesService
-import com.orange.ods.demo.domain.recipes.RecipesServiceImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -25,13 +22,9 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object DomainModule {
+object DataModule {
 
     @Singleton
     @Provides
-    fun provideDataStoreService(@ApplicationContext context: Context): DataStoreService = DataStoreServiceImpl(context)
-
-    @Singleton
-    @Provides
-    fun provideRecipesService(recipesRepository: RecipesRepository): RecipesService = RecipesServiceImpl(recipesRepository)
+    fun provideRecipesRepositoryImpl(@ApplicationContext context: Context): RecipesRepository = RecipesRepositoryImpl(context)
 }