+++ demo/src/main/java/com/orange/ods/demo/domain/datastore/DataStoreServiceImpl.kt
@@ -8,7 +8,7 @@
  * /
  */
 
-package com.orange.ods.demo.domain
+package com.orange.ods.demo.domain.datastore
 
 import android.content.Context
 import androidx.datastore.core.DataStore