+++ demo/src/main/java/com/orange/ods/demo/ui/components/bottomnavigation/ComponentBottomNavigation.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.demo.ui.components.bottomnavigation
 
-import androidx.annotation.DrawableRes
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -35,6 +34,7 @@ import com.orange.ods.demo.ui.components.bottomnavigation.ComponentBottomNavigat
 import com.orange.ods.demo.ui.components.utilities.ComponentCountRow
 import com.orange.ods.demo.ui.components.utilities.ComponentCustomizationBottomSheetScaffold
 import com.orange.ods.demo.ui.components.utilities.clickOnElement
+import com.orange.ods.demo.ui.utilities.NavigationItem
 
 private object ComponentBottomNavigation {
     const val MinNavigationItemCount = 3
@@ -45,14 +45,7 @@ private object ComponentBottomNavigation {
 @Composable
 fun ComponentBottomNavigation() {
     val context = LocalContext.current
-    val navigationItems = listOf(
-        NavigationItem("Favorites", R.drawable.ic_heart),
-        NavigationItem("Search", R.drawable.ic_search),
-        NavigationItem("Information", R.drawable.ic_info),
-        NavigationItem("Notification", R.drawable.ic_notification),
-        NavigationItem("Settings", R.drawable.ic_settings)
-    )
-
+    val navigationItems = NavigationItem.values()
     val selectedNavigationItemCount = rememberSaveable { mutableStateOf(MinNavigationItemCount) }
     val selectedNavigationItem = remember { mutableStateOf(navigationItems[0]) }
 
@@ -76,13 +69,19 @@ fun ComponentBottomNavigation() {
             OdsBottomNavigation {
                 navigationItems.take(selectedNavigationItemCount.value)
                     .forEach { navigationItem ->
+                        val label = stringResource(id = navigationItem.textResId)
                         OdsBottomNavigationItem(
-                            icon = { Icon(painter = painterResource(id = navigationItem.icon), contentDescription = null) },
-                            label = navigationItem.title,
-                            selected = selectedNavigationItem.value.title == navigationItem.title,
+                            icon = {
+                                Icon(
+                                    painter = painterResource(id = navigationItem.iconResId),
+                                    contentDescription = null
+                                )
+                            },
+                            label = label,
+                            selected = selectedNavigationItem.value.textResId == navigationItem.textResId,
                             onClick = {
                                 selectedNavigationItem.value = navigationItem
-                                clickOnElement(context, navigationItem.title)
+                                clickOnElement(context, label)
                             }
                         )
                     }
@@ -90,5 +89,3 @@ fun ComponentBottomNavigation() {
         }
     }
 }
-
-private data class NavigationItem(val title: String, @DrawableRes val icon: Int)