+++ demo/src/main/java/com/orange/ods/demo/ui/components/buttons/ButtonsContained.kt
@@ -129,7 +129,7 @@ private fun ContainedButton(
 
     OdsButton(
         modifier = if (variableWidth) modifier else modifier.fillMaxWidth(),
-        icon = if (leadingIcon) painterResource(id = R.drawable.ic_search) else null,
+        icon = if (leadingIcon) painterResource(id = R.drawable.ic_coffee) else null,
         text = stringResource(if (enabled) R.string.component_state_enabled else R.string.component_state_disabled),
         onClick = {},
         enabled = enabled,