+++ demo/src/main/java/com/orange/ods/demo/ui/components/cards/CardImageFirst.kt
@@ -17,18 +17,23 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import coil.compose.rememberAsyncImagePainter
 import com.orange.ods.compose.component.card.OdsImageFirstCard
 import com.orange.ods.demo.R
+import com.orange.ods.demo.domain.recipes.LocalRecipes
 import com.orange.ods.demo.ui.components.utilities.clickOnElement
 
 @Composable
 fun CardImageFirst(customizationState: CardCustomizationState) {
     val context = LocalContext.current
+    val recipes = LocalRecipes.current
+    val recipe = rememberSaveable { recipes.filter { it.description.isNotBlank() }.random() }
 
     with(customizationState) {
         Column(
@@ -43,10 +48,14 @@ fun CardImageFirst(customizationState: CardCustomizationState) {
             val cardContainerText = stringResource(id = R.string.component_card_element_container)
 
             OdsImageFirstCard(
-                title = stringResource(id = R.string.component_element_title),
-                image = painterResource(id = R.drawable.placeholder),
-                subtitle = if (hasSubtitle) stringResource(id = R.string.component_element_subtitle) else null,
-                text = if (hasText) stringResource(id = R.string.component_element_text_value) else null,
+                title = recipe.title,
+                image = rememberAsyncImagePainter(
+                    model = recipe.imageUrl,
+                    placeholder = painterResource(id = R.drawable.placeholder),
+                    error = painterResource(id = R.drawable.placeholder)
+                ),
+                subtitle = if (hasSubtitle) recipe.subtitle else null,
+                text = if (hasText) recipe.description else null,
                 onCardClick = if (isClickable) {
                     { clickOnElement(context, cardContainerText) }
                 } else null,