+++ demo/src/main/java/com/orange/ods/demo/ui/components/cards/CardSmall.kt
@@ -16,18 +16,23 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import coil.compose.rememberAsyncImagePainter
 import com.orange.ods.compose.component.card.OdsSmallCard
 import com.orange.ods.demo.R
+import com.orange.ods.demo.domain.recipes.LocalRecipes
 import com.orange.ods.demo.ui.components.utilities.clickOnElement
 
 @Composable
 fun CardSmall(customizationState: CardCustomizationState) {
     val context = LocalContext.current
+    val recipes = LocalRecipes.current
+    val recipe = rememberSaveable { recipes.random() }
 
     with(customizationState) {
         Row(
@@ -40,9 +45,13 @@ fun CardSmall(customizationState: CardCustomizationState) {
 
             OdsSmallCard(
                 modifier = Modifier.weight(0.5f),
-                image = painterResource(id = R.drawable.placeholder),
-                title = stringResource(id = R.string.component_element_title),
-                subtitle = if (subtitleChecked.value) stringResource(id = R.string.component_element_subtitle) else null,
+                image = rememberAsyncImagePainter(
+                    model = recipe.imageUrl,
+                    placeholder = painterResource(id = R.drawable.placeholder),
+                    error = painterResource(id = R.drawable.placeholder)
+                ),
+                title = recipe.title,
+                subtitle = if (subtitleChecked.value) recipe.subtitle else null,
                 onCardClick = if (isClickable) {
                     { clickOnElement(context, cardContainerText) }
                 } else null