+++ demo/src/main/java/com/orange/ods/demo/ui/components/tabs/ScrollableTabRow.kt
@@ -14,13 +14,12 @@ import androidx.compose.runtime.Composable
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.PagerState
 import com.orange.ods.compose.component.tab.OdsScrollableTabRow
+import com.orange.ods.demo.ui.utilities.NavigationItem
 
 @OptIn(ExperimentalPagerApi::class)
 @Composable
-fun ScrollableTabRow(tabs: List<TabItem>, pagerState: PagerState, tabIconType: MainTabsCustomizationState.TabIconType, tabTextEnabled: Boolean) {
-
+fun ScrollableTabRow(tabs: List<NavigationItem>, pagerState: PagerState, tabIconType: MainTabsCustomizationState.TabIconType, tabTextEnabled: Boolean) {
     OdsScrollableTabRow(selectedTabIndex = pagerState.currentPage) {
         Tabs(tabs = tabs, pagerState = pagerState, tabIconType = tabIconType, tabTextEnabled = tabTextEnabled)
     }
-
 }