+++ lib/src/main/java/com/orange/ods/compose/component/card/OdsTitleFirstCard.kt
@@ -107,7 +107,7 @@ fun OdsTitleFirstCard(
                 contentScale = imageContentScale,
                 modifier = Modifier
                     .fillMaxWidth()
-                    .height(dimensionResource(R.dimen.card_big_image_width))
+                    .height(dimensionResource(R.dimen.card_big_image_height))
                     .let {
                         if (imageBackgroundColor != null) it.background(imageBackgroundColor) else it
                     },